section-05/app/src/main/java/com/java/jpa/hibernate/app/repository/CourseRepository.java
package com.java.jpa.hibernate.app.repository;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Repository;
import org.springframework.transaction.annotation.Transactional;

import com.java.jpa.hibernate.app.entity.Course;

import jakarta.persistence.EntityManager;

@Repository
@Transactional
public class CourseRepository {

	@Autowired
	private EntityManager em;

	public Course findById(Long id) {
		return em.find(Course.class, id);
	}

	public void deleteById(Long id) {
		em.remove(findById(id));
	}

	public Course save(Course course) {
		if (course.getId() == null) {
			em.persist(course);
			return course;
		}
		return em.merge(course);
	}

}
